first project

+++ src/Product_Menu/Product.java
@@ -37,5 +37,5 @@ public abstract class Product {
         this.quantity++;
     }
 
-    public abstract int getCost();
+    public abstract double getCost();
 }
